Rename type_aliases.ts to constants.ts
