+++ src/iclaim_service.ts
@@ -1,4 +1,4 @@
-import {Claim, HasClaim} from './type_aliases';
+import {Claim, HasClaim} from './constants';
 
 export interface IClaimService {
   hasClaim(roles: Array<string>, claim: Claim, namespace?: string): HasClaim;
